requirements.txt
a2wsgi==1.4.0
aiohttp==3.7.4
aioredis==1.3.1
amqp==5.0.5
argh==0.26.2
argon2-cffi==20.1.0
asgiref==3.3.1
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.2.2
Automat==20.2.0
Babel==2.9.0
backcall==0.2.0
billiard==3.6.3.0
bleach==3.3.0
bravado==10.6.3
bravado-core==5.17.0
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.6
daphne==3.0.1
decorator==4.4.2
defusedxml==0.6.0
discord==1.0.1
discord.py==1.6.0
Django==3.1.7
django-cacheops==5.1
django-celery-beat==2.2.0
django-cors-headers==3.7.0
django-debug-toolbar==3.2
django-debug-toolbar-request-history==0.1.3
django-discord-connector==1.2.1
django-dynamic-preferences==1.10.1
django-esi==2.0.5
django-eveuniverse==0.7.6
django-extra-views==0.13.0
django-filter==2.4.0
django-haystack==3.0
django-model-utils==4.1.1
django-notifications-hq==1.6.0
django-oscar==3.0.2
django-oscar-accounts==2.1
django-phonenumber-field==3.0.1
django-picklefield==3.0.1
django-singleton-admin-2==1.1.0
django-tables2==2.3.4
django-timezone-field==4.1.1
django-treebeard==4.4
django-userpreferences==0.1
django-widget-tweaks==1.4.8
djangochannelsrestframework==0.2.1
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
ecdsa==0.14.1
entrypoints==0.3
factory-boy==2.12.0
Faker==7.0.1
funcy==1.15
gevent==21.1.2
greenlet==1.0.0
hiredis==1.1.0
hyperlink==21.0.0
idna==2.8
incremental==17.5.0
inflection==0.5.1
ipykernel==5.5.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itypes==1.2.0
jedi==0.18.0
Jinja2==2.11.3
jsonfield==3.1.0
jsonpointer==2.0
jsonref==0.2
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kombu==5.0.2
MarkupSafe==1.1.1
mistune==0.8.4
monotonic==1.5
msgpack==1.0.2
multidict==5.1.0
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.2.0
oauthlib==3.1.0
packaging==20.9
pandocfilters==1.4.3
parso==0.8.1
persisting-theory==0.2.1
pexpect==4.8.0
phonenumbers==8.12.20
pickleshare==0.7.5
Pillow==8.2.0
prometheus-client==0.9.0
prompt-toolkit==3.0.16
ptyprocess==0.7.0
purl==1.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.8.0
PyJWT==1.7.1
PyMySQL==1.0.2
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-crontab==2.5.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-jose==3.2.0
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.0.3
redis==3.5.3
requests==2.22.0
requests-oauthlib==1.2.0
rfc3987==1.3.8
rsa==4.7.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
Send2Trash==1.5.0
service-identity==18.1.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
strict-rfc3339==0.7
swagger-spec-validator==2.7.3
swapper==1.1.2.post1
terminado==0.9.2
testpath==0.4.4
text-unidecode==1.3
tornado==6.1
traitlets==5.0.5
Twisted==21.2.0
txaio==21.2.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.11
vine==5.0.0
watchdog==2.0.2
wcwidth==0.2.5
webcolors==1.11.1
webencodings==0.5.1
websocket-client==0.57.0
websockets==8.1
widgetsnbextension==3.5.1
yarl==1.6.3
zope.event==4.5.0
zope.interface==5.2.0
